+++ src/actions/passkey/DeleteAction.php
@@ -1,0 +1,60 @@
+<?php
+/**
+ * DeleteAction.php
+ *
+ * PHP version 8.0+
+ *
+ * @author Philippe Gaultier <pgaultier@redcat.io>
+ * @copyright 2010-2022 Redcat
+ * @license https://www.redcat.io/license license
+ * @version XXX
+ * @link https://www.redcat.io
+ * @package blackcube\admin\actions\user
+ */
+
+namespace blackcube\admin\actions\passkey;
+
+use blackcube\admin\models\Administrator;
+use blackcube\admin\models\Passkey;
+use yii\base\Action;
+use yii\web\NotFoundHttpException;
+use yii\web\Response;
+use Yii;
+
+/**
+ * Class DeleteAction
+ *
+ * @author Philippe Gaultier <pgaultier@redcat.io>
+ * @copyright 2010-2022 Redcat
+ * @license https://www.redcat.io/license license
+ * @version XXX
+ * @link https://www.redcat.io
+ * @package blackcube\admin\actions\user
+ */
+class DeleteAction extends Action
+{
+    /**
+     * @var string where to redirect
+     */
+    public $targetAction = 'account';
+
+    /**
+     * @param string $id
+     * @return string|Response
+     * @throws NotFoundHttpException
+     * @throws \yii\base\InvalidConfigException
+     */
+    public function run($id)
+    {
+        $administrator = Yii::$app->user->identity;
+        /* @var Administrator $administrator */
+        $passkey = $administrator->getPasskeys()->andWhere(['id' => $id])->one();
+        if ($passkey === null) {
+            throw new NotFoundHttpException();
+        }
+        if (Yii::$app->request->isPost) {
+            $passkey->delete();
+        }
+        return $this->controller->redirect([$this->targetAction]);
+    }
+}