+++ src/assets/webpack/src/app/attributes/index.ts
@@ -10,3 +10,5 @@ export * from './toggle-dependencies';
 export * from './fold';
 export * from './url-generator';
 export * from './view-edit';
+export * from './attach-device';
+export * from './login-device';